Adding tests
- Some things are improved to allow tests. Removing some unecessary parameters, remove some
functions and remove the coupling with the class 'Service' to avoid include use Qt Network lib
(Service is using QTcpSocket) in the simple tests.

- Adding tests to Ninjam::Server

- Allowing custom socket creation for test purpouses

- Adding tests for ninjam Server Messages

- Improving TestServer to use _data() functions

- Separating server messages handling to allow more isolated tests. A new class (ServerMessagesHandler)
was created to take care of server message handling. The Ninjam::Service class is using a instance of this
new class as a member. Now is easy test the message handling because this behavior is separated and isolated.

- Adding tests for ninjam messages using data collected by Wireshark.
  Two data set was collected. This first data set was collected connecting in a public ninjam server with 4 players.
The second data set was collected connecting in a public full ninjam server. This data set is usefull to test this
exception case (Full server).