Improving the checking for new vst plugins close #227
    Every time JamTaba is started the plugin finder dialog open and close quickly (because no new plugins are detected). Every time
Jamtaba 64 bits is started the 32 bits plugins are recognized as "new installed plugins". The same problem in Jamtaba 32 bits.
The solution was check the executable format (32 or 64 bits) in windows.
    The code to do this 'checking' is different in Windows and Mac, and was duplicated in the last version of Jamtaba code. I refatored the classes
to reuse this 'checking code'. The code is checking if a file is a valid plugin or not.
    To achieve a clean multiplatform implementation I'm using a common header (VstPluginChecker.h) and two implementation files (WindowsVstPluginChecker.cpp and MacVstPluginChecker.cpp).
The correct file for each platform (Mac or Win) is included in the project files (VstScanner.pro and Standalone.pro).