Choosing horizontal or vertical layout close #143
    The main change in this commit is related with 'how the tracks layout is builded'. The .ui files
created by QtCreator GUI tool are deleted and the layout code is writed directly in the 'BaseTrackView'
and  TrackGroupView classes. This change add more flexibility and a clean way to BaseTrackView derived
classes (LocalTrackView, NinjamTrackView) change the base layout.
    The old code was using widgets inside widgets (conteiners) to build the layout. This works, but the
'child layouts' is a clean way to achieve the same effect. Now the JamTaba code is using child layouts to
build the tracks layouts. Other places of JamTaba gui are still using widgets inside widgest and will be
refactored soon.
    Now the NinjamTrackView class has a 'orientation' member and can be 'horizontal' or 'vertical'. When
the orientation is changed the child layout orientation is changed to achieve the desired effect.
    These are the tasks in this commit:

- Removing some unecessary conteiner widgets and replacing by child layouts
- Fixing local inputs layout when collapse
- Replacing custom enums related to widgets orientation to Qt::Orientation
- Removing the files TrackGroupView.ui and BaseTrackView.ui, now these widgets are builded in the code.
- Adding buttons to change tracks layout orientation (vertical or horizontal)
- Using vertical scroll in crowded rooms with horizontal layout
- Storing and remembering last used tracks layout orientation
- Adding icons (little speakers) in level slider
- Replacing some 'magic numbers' with constants
- Using QGridlayout in BaseTrackView mainlayout