Fixing a bug in master gain volume in fresh installed JT
- The first time Jamtaba is executed (after a fresh install) the
settings file is not created yet. In this case the user preferences are
not readed from the file and the old code was using zero to the master
gain volume (the fader appears 100% to left). This can be a very bad
"first impression". Changed to use "1" (unit gain) as default master
volume in first execution.