Fixing a little mess in VST preferences dialog
The problem appears when we have no vst folders to scan (an empty list),
add one folder, and click in "Clean and Scan". The folder is removed and
the vst list is empty. This is a side effect of commit
a6cab8015cedd233c1d40bcf138da08bf19b5c6a