Fixing sound crackling in public servers stream fixed #203
- After a lot of tests I realize the problem appears when listen something in
ninbot and next listen something in ninjamer.com. Ninbot streams are stereo, and
ninjamer.com are mono streams. The streamer internal audio buffer was initialized
to stereo (ninbot) but was receiving mono data (fom ninjamer.com). The solution is
just delete this internal buffer when a stream is stopped and recreate when another
stream is started.

- Removing an unecessary construtor in NinjamRoomStreamerNode

- After discover the problem in #203 I was suspecting from something in the class
Audio::SamplesBuffer. So, I add some Unit tests. Now I know the problem is not in SamplesBuffer class,
but the tests are cool and maybe usefull in the future. So, I'm adding some tests to the SamplesBuffer.

- Drawing a progress circle when buffering the public room audio stream. This is cool and helps
a lot to understad if the buffering state is working or not.