Fixing screen initialization problems
Fixing #343. If MainWindow::showFullScreen() is called after a setMinimumSize(), and the current
minimum size is less then desktop size the fullScreen is buggy because we are forcing an
impossible minimum size (the minimum size is too large). This code is ensuring the minimum
window size is a valid window size.