Fixing screen initialization problems (#344)
* Fixing screen initialization problems

Fixing #343. If MainWindow::showFullScreen() is called after a setMinimumSize(), and the current
minimum size is less then desktop size the fullScreen is buggy because we are forcing an
impossible minimum size (the minimum size is too large). This code is ensuring the minimum
window size is a valid window size.

 - Improving the window minimumSize to avoid cropping in small monitors. Now the window topBar (title area) is subtracted from minimum window height. Now the window is always fully visible, no more cropping.

- When the window is fullscreen the topBarHeight is zero (no title window) and the window minimum size will be bigger than screen available size. The solution was 'save' the last 'topBarHeight' and use this last value when the topBarHeight is zero (full screen).