Using modified text editors in AU/VST plugins (#691)  fix #668, fix #661, fix #657
This is a AU/VST plugins change only: replacing QLineEdit (input text fields) with a QDialog containing a text editor. This QDialog is invisible, showed with same size and position of original text field. Using this trick most hosts can´t eat the keystrokes and SPACE and ENTER work ok in VST/AU plugin.

Testes in Sonar and the result is not good, I will address this soon.