Fixing VST plugin cropped screen (fix #1208, fix #1198)
New Qt versions (I`m using 5.9.9) in windows are not working well
with QWinWidget. The VST window is cropped in many hosts, except
Reaper. I tried compile Qt 5.12.9 to check if the issue is fxed,
but can`t compile. So I`m just doing a workround and shifing the VST
screen when the host is not Reaper.